+++ dbhandler/DBHandler.h
@@ -113,7 +113,7 @@ public:
 
 	bool getWaypointFromTable(WaypointModel &waypointModel, bool max);
 
-	bool getWaypointValues(int& id, float& longitude, float& latitude, int& declination, int& radius);
+	bool getWaypointValues(int& id, double& longitude, double& latitude, int& declination, int& radius);
 
 	bool insert(std::string table, std::string fields, std::string values);
